Added `bin/node-lambda` test
I divided files
